✅Logging tests working again with new folder layout.
