+++ packages/datadog_flutter_plugin/integration_test_app/integration_test/logging_test.dart
@@ -12,6 +12,7 @@ import 'common.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
+  kManualIsWeb = kIsWeb;
 
   testWidgets('test logging scenario', (WidgetTester tester) async {
     var recordedSession = await openTestScenario(tester, 'Logging Scenario');
